Добавлен адаптер для перехода на firebase
