+++ .env
@@ -1,0 +1,1 @@
+VITE_REMOTE_STORE=local
