+++ README.md
@@ -58,7 +58,7 @@
 
 - [x] Использован TypeScript.
 - [x] Низкая связанность клиентского кода с хранилищем.
-  - [storageService](src/services/storageService.ts)
-  - Клиентский код приложения не завязан жестко на конкретное хранилище данных. В данном приложении используется storageService для работы с LocalStorage, что позволяет легко заменить его на другое хранилище, например, Firebase, изменив код только в одном месте.
+  - [storageService](src/services/index.ts)
+  - Клиентский код приложения не завязан жестко на конкретное хранилище данных. В данном приложении используется `storageService`, который может быть переключен между LocalStorage и Firebase с использованием переменной окружения `REACT_APP_REMOTE_STORE`. Это позволяет легко заменить одно хранилище на другое без изменений в клиентском коде.
 - [x] Настроен CI/CD.
   - [Настройки CI/CD](.github/workflows/deploy.yml)
