+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.6",
     "dompurify": "^3.1.6",
+    "dotenv": "^16.4.5",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
